Refactored ExceptionalLineage module path, EventLog cleanup
